Add policy and confirmation to delete structure.

+++ app/Http/Controllers/CP/Structures/StructuresController.php
@@ -217,6 +217,8 @@ class StructuresController extends CpController
     {
         $structure = Structure::findByHandle($structure);
 
+        $this->authorize('delete', $structure, 'You are not authorized to delete this structure.');
+
         return Structure::delete($structure);
     }
 }
